MM-112-menuoptions: moved the hamburger unit test to header folder with hamburger button code
